# swagger.yml
swagger: '2.0'
info:
    description: 'Here you can find documentation for Smart Goals App API'
    version: 1.0.0
    title: 'Smart Goals App API Documentation'
    termsOfService: 'http://swagger.io/terms/'
    contact:
        email: maksym@gmail.com
    license:
        name: 'Apache 2.0'
        url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
basePath: /api
...
paths:
...
    /milestones/{milestoneId}:
        get:
            tags: [milestones]
            summary: 'Find milestone by id'
            description: ''
            operationId: getmilestonebyid
            produces: [application/json]
            parameters: [{name: milestoneId, in: path, description: 'Existing milestone id', required: true, type: number}]
            responses: {'200': {description: 'Successful operation', schema: {$ref: '#/definitions/Milestone'}}, '404': {description: 'Milestone not found'}}
        put:
            tags: [milestones]
            summary: 'Update a milestone'
            description: ''
            operationId: updatedmilestone
            consumes: [application/json]
            produces: [application/json]
            parameters: [{name: milestoneId, in: path, description: 'Existing milestone id', required: true, type: number}, {in: body, name: body, description: 'Milestone object with updated fields', required: true, schema: {$ref: '#/definitions/Milestone'}}]
            responses: {'204': {description: 'Milestone updated'}, '400': {description: 'Invalid input'}}
        delete:
            tags: [milestones]
            summary: 'Delete a milestone'
            description: ''
            operationId: deletemilestone
            produces: [application/json]
            parameters: [{name: milestoneId, in: path, description: 'Existing milestone id', required: true, type: number}]
            responses: {'204': {description: 'Milestone deleted'}}
...
definitions:
...
    Milestone:
        type: object
        properties:
            id: {type: number, readOnly: true}
            goalId: {type: number, readOnly: true}
            name: {type: string, format: string}
            description: {type: string, format: string}
            plannedDate: {type: string, format: date-time}
            actualDate: {type: string, format: date-time}
            value: {type: number, minimum: 1, maximum: 10}
            factors: {type: array, readOnly: true, items: {$ref: '#/definitions/Factor'}}
    Factor:
        type: object
        properties:
            id: {type: number, readOnly: true}
            name: {type: string, format: string}
            description: {type: string, format: string}
            value: {type: number, minimum: 0, maximum: 10}
            weight: {type: number, minimum: 0, maximum: 10}
            contributors: {type: array, readOnly: true, items: {$ref: '#/definitions/Contributor'}}
    Contributor:
        type: object
        properties:
            id: {type: number, readOnly: true}
            name: {type: string, format: string}
            description: {type: string, format: string}
...